Add authentication

+++ app/actions/prisma_seasons.js
@@ -2,8 +2,16 @@
 
 import { prisma } from '@/src/lib/prisma'
 import { revalidatePath } from 'next/cache'
+import { auth } from '@/auth'
 
 export async function createSeason(formData) {
+  const session = await auth()
+  if (!session?.user?.email || session.user.email !== process.env.ADMIN_EMAIL) {
+    return {
+      error: 'Vous devez être connecté pour créer une saison',
+    }
+  }
+
   try {
     const name = formData.get('name')
     await prisma.season.create({
@@ -15,10 +23,16 @@ export async function createSeason(formData) {
     }
   }
   revalidatePath('/admin')
-  revalidatePath('/location')
 }
 
 export async function updateSeason(id, newTitle) {
+  const session = await auth()
+  if (!session?.user?.email || session.user.email !== process.env.ADMIN_EMAIL) {
+    return {
+      error: 'Vous devez être connecté pour créer une saison',
+    }
+  }
+
   try {
     await prisma.season.update({
       where: { id },
@@ -27,7 +41,6 @@ export async function updateSeason(id, newTitle) {
       },
     })
     revalidatePath('/admin')
-    revalidatePath('/location')
     return { success: true }
 
   } catch (error) {
@@ -37,6 +50,13 @@ export async function updateSeason(id, newTitle) {
 }
 
 export async function deleteSeason(id) {
+  const session = await auth()
+  if (!session?.user?.email || session.user.email !== process.env.ADMIN_EMAIL) {
+    return {
+      error: 'Vous devez être connecté pour créer une saison',
+    }
+  }
+  
   try {
     await prisma.season.delete({
       where: { id },
@@ -46,5 +66,4 @@ export async function deleteSeason(id) {
     return {error: error.message}
   }
   revalidatePath('/admin')
-  revalidatePath('/location')
 }
